pack-bundle command + write callback

+++ src/util.rs
@@ -266,7 +266,7 @@ pub fn file_path_to_uri(file_path: &str) -> String {
     format!("file:///{}", path)
 }
 
-pub fn bytes_to_human_readable(bytes: u64) -> String {
+pub fn bytes_to_human_readable(bytes: u32) -> String {
     let units = ["B", "KB", "MB", "GB"];
     let mut bytes = bytes as f64;
     let mut unit = 0;
